requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo;python_version<"3.9"
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
Django==4.2.4
einops==0.6.1
faiss-cpu==1.7.4
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.9.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.279
langsmith==0.0.33
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.24.4
openai==0.28.0
packaging==23.1
Pillow==10.0.0
pydantic==2.3.0
pydantic_core==2.6.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
SQLAlchemy==2.0.20
sqlparse==0.4.4
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
matplotlib
io
